requirements.txt
flask
flask-wtf
mysql-connector-python
bcrypt
python-dotenv